seg-stacks: special fx saving prologue generation for fx passing functions
In librcd the default morestack implementation will destroy the floating point
and other extended registers so a slower form of morestack is used which saves
and restores those when context switching. Instead of using the slower variant
for all functions we detect what functions has them in their argument list or
return value, so the slower variant is only used for those special functions.