seg-stacks: fix max call frame size not being stack aligned
If the function has a call frame size it must be stack aligned as the call
lowering will pad it so the stack is aligned when the call is made. This error
caused too little space to be allocated for the stacklet when the max call frame
size was not a multiple of 16.